backend/model/train.py
"""
Training Script for LSTM Autoencoder

This script handles training the LSTM autoencoder on log sequences
for anomaly detection. Will be implemented during Phase 2.
"""

import os
import json
import torch
import torch.nn as nn
import torch.optim as optim
from torch.utils.data import DataLoader, Dataset
import numpy as np
from sklearn.metrics import precision_score, recall_score, f1_score, roc_auc_score
from typing import List, Tuple, Dict, Optional

from .lstm_autoencoder import LSTMAutoencoder, save_model, load_model
from .feature_extractor import LogTokenizer, extract_features


class LogSequenceDataset(Dataset):
    """Dataset class for log sequences."""
    
    def __init__(self, sequences: List[List[int]], labels: Optional[List[int]] = None):
        """
        Initialize dataset.
        
        Args:
            sequences: List of tokenized log sequences
            labels: Optional labels for supervised learning
        """
        self.sequences = sequences
        self.labels = labels
        
    def __len__(self):
        return len(self.sequences)
        
    def __getitem__(self, idx):
        sequence = torch.tensor(self.sequences[idx], dtype=torch.long)
        
        if self.labels is not None:
            label = torch.tensor(self.labels[idx], dtype=torch.float)
            return sequence, label
        
        return sequence


class LogAnomalyTrainer:
    """Trainer for log anomaly detection model."""
    
    def __init__(self, 
                 model: LSTMAutoencoder,
                 device: str = "cpu",
                 learning_rate: float = 0.001):
        """
        Initialize trainer.
        
        Args:
            model: LSTM autoencoder model
            device: Training device (cpu/cuda)
            learning_rate: Learning rate for optimizer
        """
        self.model = model.to(device)
        self.device = device
        self.optimizer = optim.Adam(model.parameters(), lr=learning_rate)
        self.criterion = nn.MSELoss()
        
        # Training history
        self.train_losses = []
        self.val_losses = []
        
    def train_epoch(self, dataloader: DataLoader) -> float:
        """
        Train for one epoch.
        
        Args:
            dataloader: Training data loader
            
        Returns:
            Average training loss
            
        TODO: Phase 2 implementation
        - Iterate through batches
        - Forward pass through autoencoder
        - Compute reconstruction loss
        - Backward pass and optimization
        """
        self.model.train()
        total_loss = 0.0
        num_batches = 0
        
        print("TODO: Implement train_epoch() in Phase 2")
        
        # Placeholder training loop
        for batch_idx, batch in enumerate(dataloader):
            if isinstance(batch, tuple):
                sequences, labels = batch
            else:
                sequences = batch
                
            # TODO: Implement actual training step
            loss = torch.tensor(0.1 + 0.01 * np.random.random())  # Mock loss
            total_loss += loss.item()
            num_batches += 1
            
        return total_loss / num_batches if num_batches > 0 else 0.0
        
    def validate_epoch(self, dataloader: DataLoader) -> float:
        """
        Validate for one epoch.
        
        Args:
            dataloader: Validation data loader
            
        Returns:
            Average validation loss
            
        TODO: Phase 2 implementation
        - Iterate through validation batches
        - Compute reconstruction loss without gradients
        - Return average loss
        """
        self.model.eval()
        total_loss = 0.0
        num_batches = 0
        
        print("TODO: Implement validate_epoch() in Phase 2")
        
        with torch.no_grad():
            for batch in dataloader:
                if isinstance(batch, tuple):
                    sequences, labels = batch
                else:
                    sequences = batch
                    
                # TODO: Implement actual validation step
                loss = torch.tensor(0.1 + 0.01 * np.random.random())  # Mock loss
                total_loss += loss.item()
                num_batches += 1
                
        return total_loss / num_batches if num_batches > 0 else 0.0
        
    def train(self, 
              train_loader: DataLoader,
              val_loader: DataLoader,
              num_epochs: int = 15,
              save_path: str = "artifacts/model.pth") -> Dict:
        """
        Full training loop.
        
        Args:
            train_loader: Training data loader
            val_loader: Validation data loader
            num_epochs: Number of training epochs
            save_path: Path to save trained model
            
        Returns:
            Training history dictionary
            
        TODO: Phase 2 implementation
        - Train for specified epochs
        - Monitor training and validation loss
        - Save best model based on validation loss
        - Return training metrics
        """
        print(f"TODO: Train model for {num_epochs} epochs")
        print(f"TODO: Save best model to {save_path}")
        
        best_val_loss = float('inf')
        
        for epoch in range(num_epochs):
            # Train and validate
            train_loss = self.train_epoch(train_loader)
            val_loss = self.validate_epoch(val_loader)
            
            # Record losses
            self.train_losses.append(train_loss)
            self.val_losses.append(val_loss)
            
            print(f"Epoch {epoch+1}/{num_epochs}: "
                  f"Train Loss: {train_loss:.4f}, Val Loss: {val_loss:.4f}")
            
            # Save best model
            if val_loss < best_val_loss:
                best_val_loss = val_loss
                os.makedirs(os.path.dirname(save_path), exist_ok=True)
                save_model(self.model, save_path)
                print(f"Saved best model with val_loss: {val_loss:.4f}")
                
        return {
            'train_losses': self.train_losses,
            'val_losses': self.val_losses,
            'best_val_loss': best_val_loss
        }


def determine_threshold(model: LSTMAutoencoder,
                       val_loader: DataLoader,
                       percentile: float = 95.0) -> float:
    """
    Determine anomaly detection threshold from validation set.
    
    Args:
        model: Trained autoencoder model
        val_loader: Validation data loader
        percentile: Percentile for threshold (e.g., 95th percentile)
        
    Returns:
        Anomaly detection threshold
        
    TODO: Phase 2 implementation
    - Compute reconstruction errors on validation set
    - Calculate specified percentile as threshold
    - Return threshold value
    """
    print(f"TODO: Determine threshold at {percentile}th percentile")
    
    model.eval()
    reconstruction_errors = []
    
    with torch.no_grad():
        for batch in val_loader:
            if isinstance(batch, tuple):
                sequences, labels = batch
            else:
                sequences = batch
                
            # TODO: Compute actual reconstruction errors
            # Mock reconstruction errors for now
            batch_errors = np.random.exponential(0.1, size=len(sequences))
            reconstruction_errors.extend(batch_errors)
            
    threshold = np.percentile(reconstruction_errors, percentile)
    print(f"Determined threshold: {threshold:.4f}")
    
    return threshold


def evaluate_model(model: LSTMAutoencoder,
                  test_loader: DataLoader,
                  threshold: float) -> Dict:
    """
    Evaluate model performance on test set.
    
    Args:
        model: Trained autoencoder model
        test_loader: Test data loader
        threshold: Anomaly detection threshold
        
    Returns:
        Evaluation metrics dictionary
        
    TODO: Phase 2 implementation
    - Compute reconstruction errors on test set
    - Apply threshold to get binary predictions
    - Calculate precision, recall, F1, AUC
    - Return metrics dictionary
    """
    print("TODO: Evaluate model performance in Phase 2")
    
    model.eval()
    all_scores = []
    all_labels = []
    
    with torch.no_grad():
        for batch in test_loader:
            if isinstance(batch, tuple):
                sequences, labels = batch
                all_labels.extend(labels.numpy())
            else:
                sequences = batch
                
            # TODO: Compute actual reconstruction errors
            # Mock scores for now
            batch_scores = np.random.exponential(0.1, size=len(sequences))
            all_scores.extend(batch_scores)
    
    # If we have labels, compute metrics
    if all_labels:
        predictions = (np.array(all_scores) > threshold).astype(int)
        
        precision = precision_score(all_labels, predictions)
        recall = recall_score(all_labels, predictions)
        f1 = f1_score(all_labels, predictions)
        auc = roc_auc_score(all_labels, all_scores)
        
        metrics = {
            'precision': precision,
            'recall': recall,
            'f1_score': f1,
            'roc_auc': auc,
            'threshold': threshold
        }
    else:
        metrics = {
            'threshold': threshold,
            'num_samples': len(all_scores),
            'anomaly_rate': np.mean(np.array(all_scores) > threshold)
        }
    
    return metrics


def main():
    """Main training function."""
    print("Starting LSTM Autoencoder Training...")
    
    # TODO: Phase 2 - Load and preprocess data
    print("TODO: Load datasets (Loghub, Landauer, Kaggle)")
    print("TODO: Parse logs and extract features")
    print("TODO: Create train/val/test splits")
    
    # Mock data for now
    vocab_size = 10000
    sequence_length = 20
    
    # Create model
    model = LSTMAutoencoder(
        vocab_size=vocab_size,
        embedding_dim=128,
        hidden_dim=128,
        latent_dim=64,
        num_layers=2,
        sequence_length=sequence_length
    )
    
    # Create trainer
    device = "cuda" if torch.cuda.is_available() else "cpu"
    trainer = LogAnomalyTrainer(model, device=device)
    
    print(f"Model created with {sum(p.numel() for p in model.parameters())} parameters")
    print(f"Training on device: {device}")
    
    # TODO: Create actual data loaders
    print("TODO: Create DataLoaders in Phase 2")
    print("TODO: Train model and determine threshold")
    print("TODO: Evaluate on test set and save metrics")


if __name__ == "__main__":
    main() 