backend/requirements.txt
# Core Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Kafka Integration
kafka-python==2.0.2

# Database & Search
elasticsearch==8.10.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Machine Learning
torch==2.0.1
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3

# Optional ML Libraries
transformers==4.33.2

# Data Processing
python-json-logger==2.0.7

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Monitoring
prometheus-flask-exporter==0.23.0

# Development
flake8==6.0.0
pylint==2.17.5 